move file_has_text to fixture
