+++ tests/conftest.py
@@ -1,5 +1,7 @@
 import logging
+import mmap
 import os
+import re
 from pathlib import Path
 from unittest import mock
 
@@ -326,3 +328,37 @@ def array_config_data(simulator_config_data):
 @pytest.fixture
 def shower_config_data(simulator_config_data):
     return simulator_config_data["common"] | simulator_config_data["showers"]
+
+
+@pytest.fixture
+def file_has_text(file, text):
+    """
+    Check whether a file contain a certain piece of text.
+
+    Parameters
+    ----------
+    file: str
+        Path of the file.
+    text: str
+        Piece of text to be searched for.
+
+    Returns
+    -------
+    bool
+        True if file has text.
+    """
+
+    try:
+        with open(file, "rb", 0) as string_file, mmap.mmap(
+            string_file.fileno(), 0, access=mmap.ACCESS_READ
+        ) as text_file_input:
+            re_search_1 = re.compile(f"{text}".encode())
+            search_result_1 = re_search_1.search(text_file_input)
+            if search_result_1 is None:
+                return False
+
+            return True
+    except FileNotFoundError:
+        return False
+    except ValueError:
+        return False
